Improved the code vastly.
This adds Debug implementation to all the types and Copy implementation
to all the types that can have it. It also improves the quality of the
code in various places by both using Rust 1.26 changes and having learned
from Clippy lints. In fact, this should even make the library work faster.